+++ app/utils/getConfig.js
@@ -8,16 +8,24 @@ export function getSourceName(source:PropertiesType):string {
     return sourcename[source.id] !== undefined ? sourcename[source.id] : source.title;
 }
 
-export function getLocationName(location:string):string {
-    const named_locations = window.configruntime.named_locations;
+export function getCustomLocation(location:string):Object {
     const additional_location = window.configruntime.additional_locations;
 
-    if( named_locations[location] !== undefined)
-        return named_locations[location]
     const custom_location = additional_location.find(
-        function(custom_location) {
-        return custom_location.properties.id === location;
-    });
+        function (custom_location) {
+            return custom_location.properties.id === location;
+        });
+    return custom_location;
+}
+
+export function getLocationName(location:string):string {
+    // old code, keep this for other geodashboard
+    //const named_locations = window.configruntime.named_locations;
+    //if( named_locations[location] !== undefined)
+    //    return named_locations[location]
+
+    const custom_location = getCustomLocation(location);
+
     if( custom_location)
         return custom_location.properties.title;
     return location;
@@ -35,4 +43,15 @@ export function getParameterName(parameter:string):?string {
 
 export function getTrendSettings() {
     return window.configruntime.trend_settings;
+}
+
+
+export function getColor(source: string): string {
+    let sourcecolor = window.configruntime.sourcecolor;
+    return sourcecolor[source] !== undefined ? sourcecolor[source] : '#17495B';
+}
+
+export function getTrendColor(source: string): string {
+    let trend_colors = window.configruntime.trend_colors;
+    return trend_colors[source] !== undefined ? trend_colors[source] : '#7F7F7F';
 }